test(use-window-event): migrate tests to browser mode
