--- packages/react/src/hooks/use-window-event/index.test.tsx
@@ -1,22 +1,0 @@
-import { a11y, renderHook } from "#test"
-import { useWindowEvent } from "./"
-import { Basic } from "./index.stories"
-
-describe("useWindowEvent", () => {
-  test("renders with no a11y violations", async () => {
-    await a11y(<Basic />)
-  })
-
-  test("adds and removes an event listener to the window", () => {
-    const eventType = "resize"
-    const handler = vi.fn()
-
-    const { unmount } = renderHook(() => useWindowEvent(eventType, handler))
-    window.dispatchEvent(new Event(eventType))
-    expect(handler).toHaveBeenCalledTimes(1)
-
-    unmount()
-    window.dispatchEvent(new Event(eventType))
-    expect(handler).toHaveBeenCalledTimes(1)
-  })
-})
